refactor(util): refactor functions in util directory
유틸 디렉토리를 구조화함에 따라 의존하는 파일들을 수정함.
기타 코드를 최적화함.
facebook Oauth를 제거함에 따라 facebook 라우터를 삭제함.
