+++ back/test/graphql/friend.test.js
@@ -2,7 +2,7 @@ const request = require('supertest');
 
 const graphqlPath = require('../../config/graphqlPath');
 const { app } = require('../../app');
-const signJWT = require('../../util/signJWT');
+const signJWT = require('../../util/jwt/signJWT');
 
 let token;
 beforeAll(async (done) => {
